==> Classifying Cybersecurity Incidents with Machine Learning/1.Train_&_Tests_Data.ipynb <==
# Train Data

import pandas as pd

large_file_path = r"C:\Users\Lenovo\Downloads\GUIDE_Train.csv.zip"
output_file_path = "New_train_cybersecurity.csv"

# Initializing an empty DataFrame to store the rows
rows_to_keep = pd.DataFrame()

# Reading the dataset in chunks of 100,000 rows
chunk_size = 100000
rows_needed = 100000
rows_collected = 0

for chunk in pd.read_csv(large_file_path, chunksize=chunk_size):
    if rows_collected < rows_needed:
        # Calculate remaining rows needed
        rows_to_add = rows_needed - rows_collected
        if len(chunk) <= rows_to_add:
            rows_to_keep = pd.concat([rows_to_keep, chunk])
            rows_collected += len(chunk)
        else:
            rows_to_keep = pd.concat([rows_to_keep, chunk.iloc[:rows_to_add]])
            rows_collected += rows_to_add
    else:
        break

#saving the collected rows in csv file
rows_to_keep.to_csv(output_file_path, index=False)
print(f"Saved {rows_collected} rows to {output_file_path}.")

# Test Data

large_file_path = r"C:\Users\Lenovo\Downloads\GUIDE_Test.csv.zip"
output_file_path = "New_test_cybersecurity.csv"

rows_to_keep = pd.DataFrame()

chunk_size = 100000
rows_needed = 100000
rows_collected = 0

for chunk in pd.read_csv(large_file_path, chunksize=chunk_size):
    if rows_collected < rows_needed:
        # Calculate remaining rows needed
        rows_to_add = rows_needed - rows_collected
        if len(chunk) <= rows_to_add:
            rows_to_keep = pd.concat([rows_to_keep, chunk])
            rows_collected += len(chunk)
        else:
            rows_to_keep = pd.concat([rows_to_keep, chunk.iloc[:rows_to_add]])
            rows_collected += rows_to_add
    else:
        break

rows_to_keep.to_csv(output_file_path, index=False)
print(f"Saved {rows_collected} rows to {output_file_path}.")
